Внедрение CQRS, унификация Id и новые DTO
Добавлены интерфейсы и реализации для паттернов Command и Query (CQRS): ICommand, ICommandHandler, IQuery, IQueryHandler, а также соответствующие команды и запросы для Cars, Parks, Positions, Products, Rights, Roles, Shifts и Users. Для каждой сущности реализованы команды создания и удаления, а также запросы на получение всех и по id. Добавлены DTO для логина (LoginRequestDto, LoginResponseDto) и объект-обёртка Unit для команд без результата. В Dto тип идентификатора изменён с int на long. В csproj добавлена папка Services. Исправлены и добавлены файлы для команд и запросов смен, устранены дублирующиеся и ошибочные namespace. Изменения направлены на внедрение архитектуры CQRS и унификацию идентификаторов.

+++ API/Application/Dtos/Cars/CarSimpleDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace Application.Dtos.Cars
 {
     public record CarSimpleDto(
-        int Id,
+        long Id,
         string Name
     );
 }
